refactor: :recycle: split app for testing
